v0.2
13 new countries added with recent CRTs
Fixed an issue where N2O emissions on unspecified managed soils were not included in bottom-up estimates for several countries. This required creating a new category in the LULUCF sector with the sector code "4.x." to capture the residual emissions here.
Fixed an issue where reported "Unspecified mix of HFCs and PFCs" were not included